requirements.txt
attrs==19.3.0
blinker==1.4
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
raven==6.10.0
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.16
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
webargs==8.0.1
PyJWT==2.4.0